% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/african_countries.R
\docType{data}
\name{african_countries}
\alias{african_countries}
\title{African Countries}
\format{
A tibble with 55 rows and 7 variables:
\describe{
\item{countries}{Name of the Country}
\item{region}{African Region where the Country is located}
\item{uni_count}{Number of Universities in the corresponding country}
\item{best_uni_rank}{Rank Africa of the best University of the country}
\item{gdp}{GDP of the country in 2021 (With a few exceptions for which the data is older). Source: ourworldindata.org}
\item{gdp_per_capita}{GDP per Capita of the country in 2021 (With a few exceptions for which the data is older). Source: ourworldindata.org}
\item{hdi}{Human Development Index of the country in 2021. Source: ourworldindata.org}
}
}
\usage{
african_countries
}
\description{
Database of African Countries with corresponding GDP and HDI data.
}
\keyword{datasets}
